Feature/cmapi 1

+++ database/migrations/2020_06_15_225204_create_table_has_privileges.php
@@ -15,7 +15,7 @@ class CreateTableHasPrivileges extends Migration
     {
         Schema::create('has_privileges', function (Blueprint $table) {
             $table->id();
-            $table->integer('privilege_group_id')->nullable(false);
+            $table->unsignedBigInteger('privilege_group_id')->nullable(false);
             $table->string('privilege_action')->nullable(false);
             $table->timestamps();
